refactor(udp-protocol): move scrape types to scrape module
