+++ packages/udp-protocol/src/lib.rs
@@ -24,6 +24,7 @@ pub mod connect;
 mod peer_id;
 pub mod request;
 pub mod response;
+pub mod scrape;
 
 pub use self::announce::*;
 pub use self::common::*;
@@ -31,3 +32,4 @@ pub use self::connect::*;
 pub use self::peer_id::{PeerClient, PeerId};
 pub use self::request::*;
 pub use self::response::*;
+pub use self::scrape::*;
